Validate context and params before calling worker
The worker function type name has changed from Dispatcher to
CommandFunc, and now takes a list of contexts that are defined to be
accepted by the function and can be deduced from the parameters passed
in.

So if the worker func is called it is guaranteed to have all the
parameters it advertises as having.

Closes #37