Limit number of diagnostics provided per file.
If a file is badly broken it can end up with a large number of diagnostics,
which can slow things down.

Limit the number sent to the client. This is currently hard coded to 50, but
should probably be settable via a command line option.

There is also a custom message from vscode to set this to a configuration value,
per project. It could perhaps honour that too.