Add an architecture decision record directory
See https://adr.github.io/

Add lightweight documents defining architectural decisions taken at a
point in time, so that the discussion around the decision-making
process is not lost.

To me this would include changes like deciding to do away with most of
the sub-projects, the decision to bring in IdeM and IdeGhcM, and so
on.

We do not have to retroactively document these, but it gives us a
place to capture decisions at least.