Merge master, refactor, one change.
`ContT`s `Applicative` instance captures `collectRequests`s traversing behavior.
Line 66 and 67: `to A._b` should be `A.b`, but trying to use a lens directly gives me out of scope errors. In fact, maybe `to A.b . _Just` should be a (affine, as it happens) traversal `A.b`.
Line 67 actually changes something! If the action specifies a kind we can see whether the client is capable of that. Feel free to discard this line.