Debouncing the call to diagnostics
This avoids constant recompilation while typing by only dispatching the
recompilation for thhe last event after a short timeout.

I'm personally very execited about this, since it solves one of the
biggest problems I had with hie: it kept the laptops fans spinning a
full speed.

I'm not yet sure all tests pass for this, I'd like first to open the
discussion on whether or not this is a reasonable approach.