Use legacy cabal commands, despite warnings
"cabal update" and "cabal install" will generate warnings on newer
cabal-install versions, but still work. The "cabal v1-update" and
"cabal v1-install" commands are not supported by older cabal-install.

So, until we have the next gen cabal-helper, we will have to put up
with some warnings when installing with the newer cabal-install.