Implement TypeDefinitionRequest
Works for variables, not for explicit constructors.
However, for explicit constructors, findDef works.

Implement tests for `data`, `newtype` and `type`.

For type defs, the original type definition will be found.
May be improved.

If the data type definition is not in scope, empty result will be sent.

Also add new dependency, hopefully this can be removed again