Update README.md
The suggested Aura package, `haskell-ide-engine-git`, has been flagged out of date and has not been updated in about a year (and did not build on my machine due to a missing `stack.yaml` file in the project repo). This commit instructs the reader to install `hadkell-ide-engine` instead, as this package seems to be maintained (and did build on my machine).