Do not traverse into Generated bindings when creating TypeMap (#1372)
In a module with a lot of derived type classes, such as `Cabal.LicenseId` generating the hover map was taking upwards of 2gb of memory. 

In future, we might still want information about these derived bindings so instead we might choose not to traverse into `Type` nodes or work out why the traversal was actually taking so long.