Drop stack support for GHC 8.2.2
This is necessary, because when we upgrade cabal-helper,
and transivitely cabal-plan to  version 0.5.0.0 in #1126,
the minimum cabal-version is 2.2.
However, each stack snapshot ships with its own version
of cabal-install that can not be changed.

The latest lts for ghc-8.2.2 is 11.18, which ships with
a cabal-install version of 2.0.0.1, which is too old
for cabal-plan 0.5.0.0.

However, in general, there is nothing that forces us to
drop support for GHC 8.2.2. Overriding the compiler version
should be possible, e.g. taking the snapshort 14.8 and
overriding the GHC version, but this requires us
to do manual constraint solving. Could be done if needed.