Update README.md
The windows pre-requirements are not optional but an absolute must when cloning. Well if you already have the right settings it will be a no-op just like installing `sudo apt install libicu-dev libtinfo-dev libgmp-dev` on linux will be a no-op when you already have them.

Setting the windows path limitation was not enough. Git also needs to be configured.