Demote HsImport func-tests to unit-test
Previously, we had long running func-tests for HsImport.
However, after the first code-action tests, we actually are
testing the same functionality over and over again.
Moreover, we dont have to test that HsImport works as intended
since they have a huge set of tests, anyways.

Now, we only test that our implementation of the plugin
truly maps correctly to HsImport, accelerating execution time
of func-tests as a side-effect. Hooray!