Strip RTS flags, since we cant honor them
Issue that caused it:
https://github.com/haskell/haskell-ide-engine/issues/1725

the problem is that we cant honour the `+RTS` flags, since these need to
be known when GHC is started. However, HIE is the compiler and is
started before we load any component.
Therefore, these options for the different component are unusable and we
strip them here to avoid a bug.
This commit is inspired by
https://github.com/mpickering/hie-bios/issues/102