Move finding the package for a filepath
Currently, hie-wrapper uses the cradle type to find the ghc version to
use on the project. Cabal-Helper does not only return the cradle type
but also initialises the packages of the project, which may take a long
time since it starts building dependencies. Moreover, it causes
hie-wrapper to take way longer than necessary to find the project type,
and thus, the ghc version to use on the project.
This violates the isolation of the cradle, some work happens before
loading the options for a filepath, some during it.
This commit unifies the behaviour with hie-bios:

  * All the work happens during loading the options for the given
    filepath.

It speeds up the initial start-up of hie if users use the implicit
cradle discovery mechanism.
It makes the implementation a bit less hacky.

While there are behavioural changes, nothing should change for everyday
users.